Collapse ValidationCanvas into our ManifestoCanvas wrapper
