+++ src/components/WindowSideBarCanvasPanel.js
@@ -5,7 +5,7 @@ import Typography from '@material-ui/core/Typography';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import { CanvasThumbnail } from './CanvasThumbnail';
-import { ValidationCanvas } from './ValidationCanvas';
+import ManifestoCanvas from '../lib/ManifestoCanvas';
 import { getIdAndLabelOfCanvases } from '../state/selectors';
 
 /**
@@ -35,8 +35,9 @@ export class WindowSideBarCanvasPanel extends Component {
         <List>
           {
             canvasesIdAndLabel.map((canvas, canvasIndex) => {
-              const validationCanvas = new ValidationCanvas(canvases[canvasIndex]);
-              const isValid = validationCanvas.hasValidDimensions;
+              const { width, height } = config.canvasNavigation;
+              const manifestoCanvas = new ManifestoCanvas(canvases[canvasIndex]);
+              const isValid = manifestoCanvas.hasValidDimensions;
               const onClick = () => { setCanvas(windowId, canvasIndex); }; // eslint-disable-line require-jsdoc, max-len
 
               return (
@@ -47,7 +48,7 @@ export class WindowSideBarCanvasPanel extends Component {
                     <CanvasThumbnail
                       className={classNames(classes.clickable)}
                       isValid={isValid}
-                      imageUrl={validationCanvas.thumbnail(config.canvasNavigation.height)}
+                      imageUrl={manifestoCanvas.thumbnail(config.canvasNavigation.height)}
                       onClick={onClick}
                       style={{
                         cursor: 'pointer',
