+++ src/components/index.js
@@ -10,7 +10,6 @@ export * from './NestedMenu';
 export * from './OpenSeadragonViewer';
 export * from './SanitizedHtml';
 export * from './ThumbnailNavigation';
-export * from './ValidationCanvas';
 export * from './ViewerNavigation';
 export * from './Window';
 export * from './WindowList';
